Move isolation base classes to snapshots project
