+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/AbstractIsolatableScalarValue.java
@@ -16,8 +16,8 @@
 
 package org.gradle.api.internal.changedetection.state;
 
-import org.gradle.api.internal.changedetection.state.isolation.Isolatable;
 import org.gradle.internal.Cast;
+import org.gradle.internal.isolation.Isolatable;
 
 import javax.annotation.Nullable;
