+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/ArrayValueSnapshot.java
@@ -16,9 +16,9 @@
 
 package org.gradle.api.internal.changedetection.state;
 
-import org.gradle.api.internal.changedetection.state.isolation.Isolatable;
-import org.gradle.api.internal.changedetection.state.isolation.IsolationException;
 import org.gradle.internal.hash.Hasher;
+import org.gradle.internal.isolation.Isolatable;
+import org.gradle.internal.isolation.IsolationException;
 
 import javax.annotation.Nullable;
 import java.util.Arrays;
