+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/ValueSnapshotter.java
@@ -22,16 +22,16 @@ import org.gradle.api.Named;
 import org.gradle.api.UncheckedIOException;
 import org.gradle.api.attributes.Attribute;
 import org.gradle.api.internal.attributes.AttributeDefinitionSnapshot;
-import org.gradle.api.internal.changedetection.state.isolation.Isolatable;
 import org.gradle.api.internal.changedetection.state.isolation.IsolatableEnumValueSnapshot;
-import org.gradle.api.internal.changedetection.state.isolation.IsolatableFactory;
 import org.gradle.api.internal.changedetection.state.isolation.IsolatableSerializedValueSnapshot;
 import org.gradle.api.internal.changedetection.state.isolation.IsolatableValueSnapshotStrategy;
-import org.gradle.api.internal.changedetection.state.isolation.IsolationException;
 import org.gradle.api.internal.model.NamedObjectInstantiator;
 import org.gradle.api.provider.Provider;
 import org.gradle.internal.Cast;
 import org.gradle.internal.classloader.ClassLoaderHierarchyHasher;
+import org.gradle.internal.isolation.Isolatable;
+import org.gradle.internal.isolation.IsolatableFactory;
+import org.gradle.internal.isolation.IsolationException;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
