+++ subprojects/core/src/main/java/org/gradle/internal/action/ConfigurableRule.java
@@ -16,7 +16,7 @@
 package org.gradle.internal.action;
 
 import org.gradle.api.Action;
-import org.gradle.api.internal.changedetection.state.isolation.Isolatable;
+import org.gradle.internal.isolation.Isolatable;
 
 public interface ConfigurableRule<DETAILS> {
     Class<? extends Action<DETAILS>> getRuleClass();
