+++ subprojects/core/src/main/java/org/gradle/internal/action/DefaultConfigurableRule.java
@@ -20,8 +20,8 @@ import org.gradle.api.ActionConfiguration;
 import org.gradle.api.artifacts.CacheableRule;
 import org.gradle.api.internal.DefaultActionConfiguration;
 import org.gradle.api.internal.changedetection.state.ArrayValueSnapshot;
-import org.gradle.api.internal.changedetection.state.isolation.Isolatable;
-import org.gradle.api.internal.changedetection.state.isolation.IsolatableFactory;
+import org.gradle.internal.isolation.Isolatable;
+import org.gradle.internal.isolation.IsolatableFactory;
 import org.gradle.internal.reflect.JavaReflectionUtil;
 
 import java.util.Arrays;
